Extract the Search module
