+++ lib/elastic_record/index.rb
@@ -3,6 +3,7 @@ require 'elastic_record/index/deferred'
 require 'elastic_record/index/documents'
 require 'elastic_record/index/manage'
 require 'elastic_record/index/mapping'
+require 'elastic_record/index/search'
 require 'elastic_record/index/settings'
 require 'elastic_record/index/mapping_type'
 
@@ -27,7 +28,7 @@ module ElasticRecord
   # [update_mapping]
   #   Update elastic search's mapping
   class Index
-    include Documents
+    include Documents, Search
     include Manage
     include Mapping, Settings
     include Analyze
